# book-network-client/src/app/services/models/borrowed-books-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BorrowedBooksResponse {
  authorName?: string;
  id?: number;
  isbn?: string;
  rate?: number;
  returned?: boolean;
  returnedApproved?: boolean;
  title?: string;
}