# book-network/src/main/java/com/mstramohz/book/book/BookResponse.java
package com.mstramohz.book.book;

import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class BookResponse {

    private Integer id;

    private String title;

    private String authorName;

    private String isbn;

    private String synopsis;

    private String owner;

    private byte[] cover;

    private Double rate;

    private Boolean archived;

    private Boolean shareable;

}